pyproject.toml
[tool.poetry]
name = "gdrating"
version = "0.1.0"
description = "gets company ratings from glassdoor"
authors = ["peterfraedrich <peterfraedrich@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
